Move sneak, split out activate method
